refactor(watcher): move to subdir
