+++ src/queue/importer.ts
@@ -3,7 +3,7 @@ import { spawn } from "child_process";
 import { getConfig } from "../config";
 import * as logger from "../logger";
 import { checkImageFolders, checkVideoFolders } from "./check";
-import LibraryWatcher from "./libraryWatcher";
+import LibraryWatcher from "./watch/libraryWatcher";
 import {
   getLength,
   isProcessing,
